MeasurementFnc/requirements.txt
azure-functions>=1.18
SQLAlchemy>=2
mariadb==1.1.9

